+++ src/main/kotlin/gay/asoji/innerpastels/client/imgui/ImGuiPanel.kt
@@ -1,0 +1,5 @@
+package gay.asoji.innerpastels.client.imgui
+
+interface ImGuiPanel {
+    fun render(open_: BooleanArray)
+}
