--- src/main/kotlin/gay/asoji/innerpastels/client/screens/imgui/ImGuiPanel.kt
@@ -1,5 +1,0 @@
-package gay.asoji.innerpastels.client.screens.imgui
-
-interface ImGuiPanel {
-    fun render(open_: BooleanArray)
-}
